Rename CCMPS.py -> main_gui.py
* log all exceptions
* simplify names
